Add Borrow<T> types to registry key methods
Rust's HashMap methods "get" and "contains_key"
take in a reference to a key. This is done as a
Q: Hash + Eq, where K: Borrow<Q> (the where is
kind of backwards here). This commit moves that
to the registry methods, which allow these methods
to use &str instead of Strings.