Fixed build, disabled most features.
The active container is not being set properly in initialization, so for
now most of the features don't work because we don't have a hook in. I
need to change things so that it uses a *const again, since a *mut
causes too many problems (and you can always use the parent hack to get
a mutable reference to any node.