Made tiling generic
There was too much duplicate code between horiziontal and vertical formatting,
so I made it generic by passing in lambdas of the code that actually changed.

Namely, lambdas to calcualte the new width/height, the edge case for the last
container in the list of children, and to calculate the offset before modifying
the next child's size.