Added tiling toggle test, toggle just toggles
toggle_active_horizontal now just changes the container to be either
vertical/horizontal now, it doesn't update the workspace automatically.
This makes it easier to test, and more efficient if doing large scale
operations. For now there is a slight performance cost, but that can be fixed
later with more work if it is necessary.