Added sub tiling containers
Sub containers are automatically removed when there are no children left, as
long as they are not the root container.

The root container's layout can be modified with the same keys that add new
containers.

The new keys are the following:

<alt>-v: Vertical container add, current active view is the only child
<alt>-h: Horizontal container add, current active view is the only child

Still some bugs to iron out, including a crashing one regarding parents and
children, but it is a good first start