Fixed crash when removing empty workspaces
When arranging the list of indices to be removed from the tree, an optimization
to reduce the number of index swaps in the adjacency list meant that workspaces
that were newer than their children (which is possible) would be removed first,
thereby making their root container no longer part of the tree.

To remedy this, this pre-optimization was removed, since this will cause the
iterative removal to fail.