Added right tree traversal by type, fixed refocus.
Before when deleting view the farthest left view would try to be focused. This
has been switched to be the farthest to the right. This feels slightly better,
but of course the correct way would be to remember whether to go left or right,
but for now this is fine (though might change to be a bit clever-er in the
future)