Added switch workspace when not focused hack
This feature isn't a hack, but the implementation is. This should be
properly implemented once we have a concept of "paths" to the active
container. However right now we just grab the workspace that reported
the last activity. It might be buggy, but if it is it's fine because it
will be changed by v0.4.0