Can move floating windows.
Right now it's very inefficient, because it actually locks the entire
tree everytime the pointer moves! So I need to move tho lock outside of
the tree so that we can test it without having to lock the tree.

Hmm wonder if I can do it with an Rc and a weak? Or maybe that's a bad
design