Added floating views, fixed loads of things
Sorry, this was a HHHHUUUGGGEEE commit.

I changed quite a few things, left a few todos still left to do...
But the resizing for floating works, and I'm mostly happy with the
outcome, considering what I had to work with.

Still panics when trying to resize a non-floating view, that is by
design