Bad code, barely working, but it kind of does
Just need to fix it up so it works properly (handles all the cases,
including resizing containers not just views and proprogating the
changes).

Consider throwing a `layout()` after each resize. It might become
expensive though, so as long as we can guarantee that the results will
be fine it should suffice...

Also *really* need to clean up the code, there are some bad hacks and
TODOs littering the place, and I'm trying to make the tree api a little
nicer...so I need to start using UUIDs everywhere I can, and you can see
it reflected in this code.