More efficient, purged logging, -useless scale var
Still buggy, there are still issues with it overflowing when multiple
children are present, and you can't resize sub views orthogonal to
their parent's container. Will need to review the algorithm to
understand how to incorporate that into the code. Will probably need
some way to recursively call up if an Err was returned from the helper.