Added fullscreen_c vec to workspace
Keeps track of all the children of the workspace that are fullscreen,
this allows us to limit what the user can do in an efficient way.

Note that there are no tests added by this, and though I'm fairly sure
this invarient is held up currently that may not always be the case.
Care must be taken to use this value elsewhere.