Added effective geometry to view, fix gap resizing
This somewhat complicates the view geometry story, because now we have
an "effective" geometry that we store that, for the most part, is in
line with what the actual geometry is. We use the effective geometry for
most calculations though (e.g tiling, resizing, etc) because otherwise
we are ignoring the gap/border that is around the window.

If some operation needs the actual geometry, they should call
"handle.get_geometry()" directly, as evidentally they need to bypass the
abstraction that the tree/container provides