Starting the Lua thread prompts the tree to tile
This allows reloading configuration options that affect the tree, e.g
gap size and eventually borders.

It would be nice to put this in the restart response, but then the new
value will not have been read in yet. So instead, everytime the Lua
thread starts it causes a re-tile. For init, this has a negilible effect
on performance because there's nothing to tile.