view_created now blocks locking the tree (fix bug)
Fixes the issue where the background would "mysteriously" crash the
program when changing the border size w/ floating windows.

The tree would not block locking, and it would try to focus on the
floating view causing the it to seize because you can't lock it multiple
times and we just swallow the error.