Stubbed out rendering title bars.
Not quite correct. We aren't handling the first case properly when
tiled, and we have that horrible gap at the bottom. We need to change
how the "add_borders" function in layout.rs works, need to calculate at
a container level, and abstract it enough that borders can tell us what
geometry we should set (so we don't have to keep adding annoying things
to lookup)