Added basic multi-head output. "Works", needs work
I also fixed the "descendant_with_handle" function, so
that it takes a generic handle and searches the right path.

Some of the issues I need to fix:
* Tiles based on everything being added to the first output, need to add
a check when making a new view on which output it's on, and then add it
to the appropriate node in the tree.
* Not tiling from the root container always (such as when switching
workspaces), so some views don't update until you move the mouse.
* New output is __init__ with a workspace == "1", should really not do
that and try to come up with a better default altogether...because
having a default of "1" is bad in general, breaks totally with a new
output (should be the "next" one, whatever that is. Specify in the
configuration file maybe?)
* Background doesn't work on that output, should look into fixing.
* There are probably other hard-coded "only do thing on output 1"
littering the code base, should hunt them down and fix them.