"Fixed" test, but the issue remains.
The issue is this:

When making a new output, what do you init the new workspace name to?
Right now we just do the "numerical" value of the output (The
`WlcOutput` value). This is a gross hack, and assumes many things:
1) That the workspace doesn't exist (e.g hotplugging will probably cause
a crash in the current state)
2) That the user binds their workspaces numerically
3) That the output number will be a number that the user has bound (e.g,
with the default setup anything >= 10 will not work)

This is a somewhat complicated issue to solve however, and will need
user-configuration level support that requires more thought. So I'm
going to leave it in it's current state for now, and fix it before 1.0.

Again, because this requires user configuration changes, I'm going to
hold of until I find the best way to expose this.