Can't construct object unless Objectable
You could still trigger the assert to fail, but then you'll just get a
runtime error rather than a bug.

As of right now, if you get an Object back from Lua you need to do a
concrete cast before you can mess with the data inside. Which is good :)