Properties are now stored and accessed correctly
This will be documented elsewhere, but it's a little crazy...

Object is called with an index it doesn't recognize (e.g button on
button object). It doesn't know what that is, so it asks its class if
there is a method defined that is bound to that name (e.g "button" has
an index method).

It gets the class because all objects store a pointer to the table
object in its table. It gets that table, and then checks the property
list on the class table for the function. Then it calls that method with
the object.

Makes sense, but the mechanism can be a bit tricky to understand beneath
all of the lifetimes and what not.