Objectable deals with raw allocation
It makes it a bit awkward to augment values when passed in with __call,
but that will just be setters on the value I think. It should be ok,
though all of the copying of the table from the setters might get
inefficient...we'll deal with it later, this is "ok" and can be
optimized.