Gahh why does this give me a func
Everything else gives me table! Even the repl from rlua! There must be a
bug somewhere. It's insane, even with the most minimal environment set
up in the repl (package.path, the string in the awesome thing, and the
fake debug) it gives me back a table in the repl

But when I run way cooler loading the rc.lua, it gives me a function and
crashes!

UPDATE: Issue has been found to be in rlua due to nested pcalls.

Reported: https://github.com/chucklefish/rlua/issues/53