Fixed user object creation
We weren't wrapping it in a table, like is expected of in Lua because
Awesome does it.

HOWEVER, they use lua_setuservalue. We don't have access to that, so we
just put it in a table. Need to double check this does the right
thing...because this doesn't seem right :(