Unexport the LUA global, provide run_with_lua instead
Future commits will require that all accesses to the Lua state will have
to be from within a single thread to avoid deadlocks. Thus, the LUA
global cannot be public any more.

Instead, this commit adds a run_with_lua() function which gets a closure
that will execute in the Lua thread and get the Lua state as its first
argument.

This commit does not yet execute the closure in the Lua thread.

Signed-off-by: Uli Schlachter <psychon@znc.in>