Override Lua's type() function
In awesome, type() of a screen returns "screen". To make this work,
awesome replace's Lua's type() function with its own version. This
commit does the same.

Note that I am not quite sure where to do this and for now I chose
rust_interop since it was convenient (there is a function that is run on
Lua state initialisation).

Signed-off-by: Uli Schlachter <psychon@znc.in>