Use a mpsc::channel for talking to the Lua thread again
The existing infrastructure of using a glib idle callback has the
downside that it silently fails when the thread running the glib main
loop panicked. The idle callbacks just sit in a queue in glib and are
not processed. This can easily lead to deadlocks where a thread waits
for its message to the Lua thread to be processed, which never happens.

To fix this, this commit re-introduces a std::sync::mpsc::channel that
is used for talking to the Lua thread. Messages are sent to the channel
and then an idle callback is scheduled to process messages on the
channel.

To actually make this work nicely when the Lua thread panics, a guard is
introduced. In its Drop implementation, it drops the receiving end of
the channel. This means that all messages that are still in flight will
be dropped as well, thus causing suitable errors for all threads that
wait for a message to be processed. Dropping the receiver also ensures
that the sender cannot be used to send further messages.

Fixes: https://github.com/way-cooler/way-cooler/issues/477
Signed-off-by: Uli Schlachter <psychon@znc.in>