Handle constructor object arguments
Doing e.g. drawin{ x = 42 } is supposed to create a drawin with its x
property set to 42. This commit actually implements this for all
supported objects except for buttons.

For buttons this causes an error during loading of the awesome default
config and so the argument to buttons is still ignored for now.

  cause: RuntimeError("Lock is an invalid modifier")

Signed-off-by: Uli Schlachter <psychon@znc.in>