Fix drawin resizing their drawable incorrectly
In a drawin's resize() method, the function update_drawing() (which
resizes the drawable) was called before the new state of the drawin was
set. This meant that the drawable was resized to the old size of the
drawin and not the new one.

Fix this by swapping the two calls.

Signed-off-by: Uli Schlachter <psychon@znc.in>