button: Use get_object_mut() instead of set_state()
There is way less copying going on when modifying the state directly
instead of having to query and then set it again.

Plus, if something tries to borrow the same object mutably twice, then
that results in an error. That is better than the old behaviour were a
second set_state() would just silently overwrite the changes done by a
first set_state().

Signed-off-by: Uli Schlachter <psychon@znc.in>