awesome.set_preferred_icon_size: Modify state in place
The old function tried to replace the "awesome" global and it did so in
a broken way. Instead, this just borrows the userdatum mutably and
modifies it in place.

Tested via adding the following in this function and checking that it
prints the expected old value when changing the preferred icon size
twice in a row:

  println!("Old preferred was {}, now {}",
  awesome_state.preferred_icon_size, val);

Fixes: https://github.com/way-cooler/way-cooler/issues/501
Signed-off-by: Uli Schlachter <psychon@znc.in>