load_surface_from_pixbuf: (Likely) work on big endian
GdkPixbuf defines its pixel format based on bytes: The first byte is the
read component, the following is green, etc. Cairo however defines its
pixel format based on uint32_t and talks about the "upper bits". Thus,
this is endian-dependent.

Thus, this code has to "write things" in a different order on big endian
machines. This is exactly what this patch does.

Note that I do not have access to a big endian machine to test this. I
am just assuming that it works.

Signed-off-by: Uli Schlachter <psychon@znc.in>