load_surface_from_pixbuf: Eliminate second index variable
As one of the previous commits showed, having two index variables is
just an invitation for mixing them up and using the wrong one. Thus,
this patch eliminates one of them. I expect, but have not checked, the
compiler to be smart enough so that this does not actually change the
generated code.

Signed-off-by: Uli Schlachter <psychon@znc.in>