Lua: Implement awesome.pixbuf_to_surface()
This is a new function that was added to awesome in commit ee944f4da83.
It gets a pointer to a GdkPixbuf instance as its argument and returns a
cairo image surface with the same content. It works similar to the
existing awesome.load_image(), but without loading a pixbuf from a file,
thus it is more general.

Fixes: https://github.com/way-cooler/way-cooler/issues/545
Signed-off-by: Uli Schlachter <psychon@znc.in>