root.keys(): Make a copy when setting keys
Before this commit, the code just kept a reference to the passed-in
table. However, Lua could easily keep a reference to this table as well
and modify it after the fact. Fix this by copying the table.

This makes something like the following work, i.e. key2 is not saved:

   local keys = { key1 }
   root.keys(keys)
   keys[2] = key2

Signed-off-by: Uli Schlachter <psychon@znc.in>