Fix XDG_DATA_DIRS and XDG_CONFIG_DIRS handling
Previously, each of these environment variables was treated as the path
to a directory.

According to the XDG Base Directory Specification, however, they are
colon separated lists of directories. To correctly handling such lists
of length > 1, we have to split them apart at the colon and process each
element separately.