Don't have a global OUTPUTS list
We already have a global screens list, which is the one that matters to
Awesome.

We still have the cached output state per output, but it's only
accessible in the Lua code.

As well, when an update occurs to the output it will automatically fetch
the screen that's associated with that output and update its metadata.