Drawables own an xdg_toplevel on creation
They don't actually draw anything to it, but got this working.

Also fixed a _massive_ UB bug where we were dropping our reference
counted wl_display and this was causing weird issues.

So now that works, which is nice.

Also this is the first patch that no longer works on Way Cooler v0.8, so
to test this fully we need to use `make` in the top directory