Added stable XDG shell support for Way Cooler
Now Awesome should work with Way Cooler without crashing ;).

The code has lots of duplicate paths, it needs to be decided how to deal
with that. It _should_ stay the same, but it may not.

In case it does stay the same, we should unify the paths and regularly
check that clients are still using the unstable v6 version.