Added test drawables that work
So this works, and fairly well at that which is nice.

Going to iterate on this, mainly by making the code less shit at first
and then finally actually rendering the contents of the drawables.

One issue that needs to be addressed is to render the drawable contents
without unnecessary copying. I'd rather not have to do that, but it
might be necessary.

The only real requirement is getting an fd to a block of memory, (i.e we
need something that implements
https://doc.rust-lang.org/nightly/std/os/unix/io/trait.AsRawFd.html)

However we _need_ to use a cairo ImageSurface since that's how the Lua
code writes out the bytes. We could use create_for_data, but that will
require using unsafe code unless we want it owned by the surface (which
while safe isn't what we want since we need some way to access the data
behind it using an fd).

Ultimately we might just copy the bytes anyways.