Add man pages
Add man and html targets to the makefile, responsible for generating two versions of the content of the manpages directory, one for man per se and one for the web.
The man pages are a basic introduction to the window manager, but I feel they should be developed further.
They are fairly short, as only one flag is supported and the specifications are already done by Awesome.
They are traduced in french and english.