Implement the clients property for tag, stored in an associate value.
Implement the public API for the client property, except for the signals.
To avoid the Rust-side cloning, values are stored as-is in an associate table (`__clients`) on the object.
This allows the clients to be shared amongst tags as the same Lua-side tables.
A bottleneck occurs when setting the new clients, as one needs to iterate over the old and new client lists to check for common values.
This should not be a performance problem though, as it is not used frequently.