Make `Object` a generic type rather than a type wrapped inside classes
By using a `PhantomData` on the `Object` type, one can then make the different awesome objects directly as `Object`s which accept different states
This remove the need to copy the traits definition in all objects (even if the macro call makes it easy) and remove the need for the Objectable trait.
This also means that most of the AnyUserData function parameters could be replaced by the concrete type using a common `FromLua` implementation, thus making the code more DRY by removing the need for explicit casting.