Remove all the warnings
With all the methods that aren't used, the dead code warnings polute the output of compilation and obfuscate the valuable warnings during development.
This might be a bit controversial, but it it will be easier to grep for the different `allow` annotations after version 1.0 than to fix a lot of warnings which should be caught during the process.