Disallow not using values and upgrade logging infrastructure
Add `unused_results` to `deny`.
Fix various warnings related to it, preferably by a code refactoring, some only falling back to `let = `.
Because a lot of warnings appeared around the log creation, I updated the dependency and redesigned a bit the code.
Nothing used the returns from the `build` command nor from `init` and it was a major source of warnings, so I removed them