"awesome' -> "client"
This will make it much easier to understand what the fuck is going on
now for both new comers and myself.

Use way-cooler for the compositor binary name still.

This way we don't need a unifying script to launch the two processes,
instead way-cooler will have an argument to spawn a client program (like
tinywl) and that is how the client can be spawned.

It should also be usable as a standalone compositor. So it should be
somewhere between rootston and tinywl. E.g. no server-side borders.

It should be able to basically do things without awesome working correctly.