Update rlua: 0.12 -> 0.16
Use feature system-lua since the builtin runtime does not support lua-lgi.

From rlua 0.16 changelog:
- API incompatible change: remove str assumptions from the remaining
places: the UserData trait and named registry keys
- HUGE API incompatible change: move most of the Lua API into Context and
require context callbacks to generate a branding lifetime, and use this
branding lifetime on handle types. Fixes a hard to trigger soundness
issue, and removes the only remaining API panic (that is not a bug). See
the documentation for the Lua type for more details.