Give startup client the wl_client connection
This will allow us to implement security going forward (i.e. only allow
that client to use those interfaces) and for us to handle client failure
more gracefully.

We don't actually do that, but this has the setup for it. It requires
more thought on how to do the latter gracefully and the former without
impacting development / debugging